+++ NEWS.md
@@ -1,5 +1,6 @@
 # plm 2.2-0
-* (tbd)
+* lagt is changed so that it can deal with time factors for which
+  cannot be coerced to numeric (ex "1950-54", "1955-59", ...)
 
 # plm 2.1-0
 * problems with vignettes (all the text in italics fixed)